# src/main/resources/swagger/api.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1-SNAPSHOT",
    "title": "atlas",
    "x-appId": "1000001128",
    "x-serviceName": "atlas"
  },
  "host": "localhost:8081",
  "basePath": "/",
...
  "paths": {
...
    "/web/roles": {
      "post": {
        "tags": [
          "role-controller"
        ],
        "summary": "addRole",
        "operationId": "addRoleUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "roleDto",
            "description": "roleDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RoleDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«RoleDto»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "role-controller"
        ],
        "summary": "updateRole",
        "operationId": "updateRoleUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "roleDto",
            "description": "roleDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RoleDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«RoleDto»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/web/roles/all": {
      "get": {
        "tags": [
          "role-controller"
        ],
        "summary": "getAllRoles",
        "operationId": "getAllRolesUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«List«RoleDto»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/web/roles/page": {
      "get": {
        "tags": [
          "role-controller"
        ],
        "summary": "getRoleByPage",
        "operationId": "getRoleByPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "size",
            "in": "query",
            "description": "size",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«PageDto«RoleDto»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/web/roles/{roleId}": {
      "get": {
        "tags": [
          "role-controller"
        ],
        "summary": "getRoleById",
        "operationId": "getRoleByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "description": "roleId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«RoleDto»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "role-controller"
        ],
        "summary": "deleteRole",
        "operationId": "deleteRoleUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "description": "roleId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«string»"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/web/userroles": {
      "put": {
        "tags": [
          "user-role-controller"
        ],
        "summary": "updateUserRole",
        "operationId": "updateUserRoleUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "userRoleDto",
            "description": "userRoleDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserRoleDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«UserRoleDto»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/web/userroles/all": {
      "get": {
        "tags": [
          "user-role-controller"
        ],
        "summary": "getAllUsers",
        "operationId": "getAllUsersUsingGET_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«List«UserRoleDto»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/web/userroles/name/{userName}": {
      "get": {
        "tags": [
          "user-role-controller"
        ],
        "summary": "getRolesByUserName",
        "operationId": "getRolesByUserNameUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "userName",
            "in": "path",
            "description": "userName",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«List«string»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/web/userroles/page": {
      "get": {
        "tags": [
          "user-role-controller"
        ],
        "summary": "getUserRoleByPage",
        "operationId": "getUserRoleByPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "size",
            "in": "query",
            "description": "size",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«PageDto«UserRoleDto»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/web/userroles/roles/{userWorkNumber}": {
      "get": {
        "tags": [
          "user-role-controller"
        ],
        "summary": "getRolesByUserId",
        "operationId": "getRolesByUserIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "userWorkNumber",
            "in": "path",
            "description": "userWorkNumber",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«List«RoleDto»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/web/userroles/{userRoleId}": {
      "get": {
        "tags": [
          "user-role-controller"
        ],
        "summary": "getUserRoleById",
        "operationId": "getUserRoleByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "userRoleId",
            "in": "path",
            "description": "userRoleId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«UserRoleDto»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "user-role-controller"
        ],
        "summary": "deleteUserRole",
        "operationId": "deleteUserRoleUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "userRoleId",
            "in": "path",
            "description": "userRoleId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«string»"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/web/userroles/{userWorkNumber}": {
      "post": {
        "tags": [
          "user-role-controller"
        ],
        "summary": "addUserRole",
        "operationId": "addUserRoleUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "userWorkNumber",
            "in": "path",
            "description": "userWorkNumber",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "roleIds",
            "description": "roleIds",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/web/users": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "addUser",
        "operationId": "addUserUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "userDto",
            "description": "userDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "user-controller"
        ],
        "summary": "updateUser",
        "operationId": "updateUserUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "userDto",
            "description": "userDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«UserDto»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/web/users/all": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "getAllUsers",
        "operationId": "getAllUsersUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«List«UserDto»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/web/users/page": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "getUserByPage",
        "operationId": "getUserByPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "size",
            "in": "query",
            "description": "size",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«PageDto«UserDto»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/web/users/{userId}": {
      "delete": {
        "tags": [
          "user-controller"
        ],
        "summary": "deleteUser",
        "operationId": "deleteUserUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "userId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«string»"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "OrgDto": {
      "type": "object",
      "properties": {
        "active": {
          "type": "boolean"
        },
        "description": {
          "type": "string"
        },
        "extensions": {
          "type": "object"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "insertBy": {
          "type": "string"
        },
        "insertTime": {
          "type": "string",
          "format": "date-time"
        },
        "name": {
          "type": "string"
        },
        "orgCode": {
          "type": "string"
        },
        "parentOrgId": {
          "type": "integer",
          "format": "int64"
        },
        "updateBy": {
          "type": "string"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time"
        },
        "userDto": {
          "$ref": "#/definitions/UserDto"
        },
        "userWorkNumber": {
          "type": "string"
        }
      }
    },
...
    "PageDto«RoleDto»": {
      "type": "object",
      "properties": {
        "content": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/RoleDto"
          }
        },
        "first": {
          "type": "boolean"
        },
        "last": {
          "type": "boolean"
        },
        "number": {
          "type": "integer",
          "format": "int32"
        },
        "numberOfElements": {
          "type": "integer",
          "format": "int32"
        },
        "size": {
          "type": "integer",
          "format": "int32"
        },
        "totalElements": {
          "type": "integer",
          "format": "int64"
        },
        "totalPages": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "PageDto«UserDto»": {
      "type": "object",
      "properties": {
        "content": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserDto"
          }
        },
        "first": {
          "type": "boolean"
        },
        "last": {
          "type": "boolean"
        },
        "number": {
          "type": "integer",
          "format": "int32"
        },
        "numberOfElements": {
          "type": "integer",
          "format": "int32"
        },
        "size": {
          "type": "integer",
          "format": "int32"
        },
        "totalElements": {
          "type": "integer",
          "format": "int64"
        },
        "totalPages": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "PageDto«UserRoleDto»": {
      "type": "object",
      "properties": {
        "content": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserRoleDto"
          }
        },
        "first": {
          "type": "boolean"
        },
        "last": {
          "type": "boolean"
        },
        "number": {
          "type": "integer",
          "format": "int32"
        },
        "numberOfElements": {
          "type": "integer",
          "format": "int32"
        },
        "size": {
          "type": "integer",
          "format": "int32"
        },
        "totalElements": {
          "type": "integer",
          "format": "int64"
        },
        "totalPages": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "Response«List«RoleDto»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/RoleDto"
          }
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "Response«List«UserDto»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserDto"
          }
        },
        "message": {
          "type": "string"
        }
      }
    },
    "Response«List«UserRoleDto»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserRoleDto"
          }
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "Response«List«string»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "Response«PageDto«RoleDto»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "$ref": "#/definitions/PageDto«RoleDto»"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "Response«PageDto«UserDto»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "$ref": "#/definitions/PageDto«UserDto»"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "Response«PageDto«UserRoleDto»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "$ref": "#/definitions/PageDto«UserRoleDto»"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "Response«RoleDto»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "$ref": "#/definitions/RoleDto"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "Response«UserDto»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "$ref": "#/definitions/UserDto"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "Response«UserRoleDto»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "$ref": "#/definitions/UserRoleDto"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "Response«string»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "RoleDto": {
      "type": "object",
      "properties": {
        "active": {
          "type": "boolean"
        },
        "description": {
          "type": "string"
        },
        "extensions": {
          "type": "object"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "insertBy": {
          "type": "string"
        },
        "insertTime": {
          "type": "string",
          "format": "date-time"
        },
        "name": {
          "type": "string"
        },
        "updateBy": {
          "type": "string"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "UserDto": {
      "type": "object",
      "properties": {
        "active": {
          "type": "boolean"
        },
        "email": {
          "type": "string"
        },
        "extensions": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "insertBy": {
          "type": "string"
        },
        "insertTime": {
          "type": "string",
          "format": "date-time"
        },
        "lastVisitAt": {
          "type": "string",
          "format": "date-time"
        },
        "ldapInsertTime": {
          "type": "string",
          "format": "date-time"
        },
        "ldapUpdateTime": {
          "type": "string",
          "format": "date-time"
        },
        "orgDto": {
          "$ref": "#/definitions/OrgDto"
        },
        "orgId": {
          "type": "integer",
          "format": "int64"
        },
        "realName": {
          "type": "string"
        },
        "roles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/RoleDto"
          }
        },
        "source": {
          "type": "string"
        },
        "updateBy": {
          "type": "string"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time"
        },
        "userName": {
          "type": "string"
        },
        "workNumber": {
          "type": "string"
        }
      }
    },
...
    "UserRoleDto": {
      "type": "object",
      "properties": {
        "active": {
          "type": "boolean"
        },
        "description": {
          "type": "string"
        },
        "extensions": {
          "type": "object"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "insertBy": {
          "type": "string"
        },
        "insertTime": {
          "type": "string",
          "format": "date-time"
        },
        "roleId": {
          "type": "integer",
          "format": "int64"
        },
        "updateBy": {
          "type": "string"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time"
        },
        "userWorkNumber": {
          "type": "string"
        }
      }
    },
...
  }
}